feat(basecamp): implement draft saving logic and dynamic edit routing Why: - To allow users to persist their drafts and continue editing them later. - Need a flexible way to store platform-specific data like Zenn's slug.
What:
- Add DB migration for 'metadata' JSONB column in 'sitecue_drafts'.
- Create reusable 'DraftEditor' component with Supabase insert/update logic.
- Refactor '/draft/new' to use the new component.
- Create dynamic route '/draft/[id]' to fetch and edit existing drafts.
